Use proxy wiring for future event buffer
